+++ src/main/builders/SectionBuilder.js
@@ -23,15 +23,17 @@ class SectionBuilder {
     // Create build module
     section.build = BuildFactory.fromSection(section)
 
+    // Build Locales
+    section.build.locales = LocaleUtils.buildLocales(section.name, section.locales, section.schema?.locales)
+    section.build.schemaLocales = LocaleUtils.buildLocales(section.name, section.schemaLocales)
+
     // Build Section Schema (this includes previously collated locales through factory methods
-    const snippetsSchema = RecursiveRenderUtils.getSnippetsSchema(section.renders)
-    section.build.schema = SectionSchemaUtils.merge(section.schema, snippetsSchema)
+    const snippetsSchema = RecursiveRenderUtils.getSnippetsBuildSchema(section.renders)
+    if (section.schema || snippetsSchema) {
+      section.build.schema = SectionSchemaUtils.build(section.schema, snippetsSchema)
+    }
 
     section.build.liquidCode = await this.buildLiquid(section.liquidCode, section.build.schema)
-
-    // Assemble Schema Locales
-    const renderSchemaLocales = RecursiveRenderUtils.getSnippetsSchemaLocales(section.renders)
-    section.build.schemaLocales = this.buildSchemaLocales(section.name, section.schemaLocales, renderSchemaLocales)
   }
 
   /**
@@ -50,31 +52,6 @@ class SectionBuilder {
     }
 
     return buildLiquidCode
-  }
-
-  /**
-   *
-   * @param {string} sectionName
-   * @param {Object} [sectionSchemaLocales={}]
-   * @param {Object} [snippetsSchemaLocales={}]
-   * @return {Object}
-   */
-  static buildSchemaLocales (sectionName, sectionSchemaLocales = {}, snippetsSchemaLocales = {}) {
-    let buildSchemaLocales = {}
-
-    buildSchemaLocales = merge(buildSchemaLocales, sectionSchemaLocales)
-    buildSchemaLocales = merge(buildSchemaLocales, snippetsSchemaLocales)
-
-    // Put Schema Locales in the appropriate section namespace
-    for (const locale of Object.keys(buildSchemaLocales)) {
-      buildSchemaLocales[locale] = {
-        sections: {
-          [sectionName]: buildSchemaLocales[locale]
-        }
-      }
-    }
-
-    return buildSchemaLocales
   }
 
   /**
@@ -187,15 +164,34 @@ class SectionBuilder {
     }
 
     const { liquidFilesWritePromise } = RecursiveRenderUtils.getSnippetsLiquidFilesWritePromise(section.renders, section.build.snippetsFolder)
-
-    return Promise.all([
-      LocaleUtils.writeSchemaLocales(section.build.schemaLocales, section.build.localesFolder),
+    const filesWritePromises = [
+      LocaleUtils.writeLocales(this.assembleLocales(section.build.locales, section.renders), section.build.localesFolder),
+      LocaleUtils.writeLocales(this.assembleLocales(section.build.schemaLocales, section.renders, true), section.build.localesFolder, true),
       FileUtils.copyFilesToFolder(assetFiles, section.build.assetsFolder),
-      FileUtils.writeFile(section.build.stylesheet, section.build.styles),
       FileUtils.writeFile(section.build.liquidFile, section.build.liquidCode),
-      FileUtils.writeFile(section.build.settingsSchemaFile, JSON.stringify(section.build.settingsSchema, null, 2)),
       liquidFilesWritePromise
-    ])
+    ]
+
+    if (section.build.styles) {
+      filesWritePromises.push(FileUtils.writeFile(section.build.stylesheet, section.build.styles))
+    }
+    if (section.build.settingsSchema?.length) {
+      filesWritePromises.push(FileUtils.writeFile(section.build.settingsSchemaFile, JSON.stringify(section.build.settingsSchema, null, 2)))
+    }
+
+    return Promise.all(filesWritePromises)
+  }
+
+  /**
+   * Assemble Locales
+   * @param {Object} sectionLocales
+   * @param {Render[]} renders
+   * @param {boolean} [isSchemaLocales=false] Defaults to storefront locales
+   * @returns {Object}
+   */
+  static assembleLocales (sectionLocales, renders, isSchemaLocales = false) {
+    const renderLocales = RecursiveRenderUtils.getSnippetsBuildLocales(renders, isSchemaLocales)
+    return merge(sectionLocales, renderLocales)
   }
 }