+++ src/main/builders/SnippetBuilder.js
@@ -1,3 +1,5 @@
+import LocaleUtils from '../../utils/LocaleUtils.js'
+import SectionSchemaUtils from '../../utils/SectionSchemaUtils.js'
 import SnippetBuild from '../models/SnippetBuild.js'
 import SvgProcessor from '../processors/SvgProcessor.js'
 
@@ -9,6 +11,17 @@ class SnippetBuilder {
    */
   static async build (snippet) {
     snippet.build = new SnippetBuild()
+
+    // Build Locales
+    snippet.build.locales = LocaleUtils.buildLocales(snippet.name, snippet.locales, snippet.schema?.locales, true)
+    snippet.build.schemaLocales = LocaleUtils.buildLocales(snippet.name, snippet.schemaLocales, null, true)
+
+    // Build Schema
+    if (snippet.schema) {
+      snippet.build.schema = SectionSchemaUtils.build(snippet.schema)
+    }
+
+    // Build Liquid Code
     snippet.build.liquidCode = await this.buildLiquid(snippet.name, snippet.liquidCode)
 
     return snippet
@@ -24,6 +37,7 @@ class SnippetBuilder {
     if (snippetName.startsWith('icon-') || snippetName.endsWith('-svg') || snippetName.endsWith('.svg')) {
       return SvgProcessor.buildSvg(snippetName, snippetLiquidCode)
     }
+
     return snippetLiquidCode
   }
 }