+++ src/main/models/abstract/ComponentBuild.js
@@ -10,6 +10,9 @@ class ComponentBuild {
   /** @type {string} **/
   #javascriptFile
 
+  /** @type {Object} **/
+  #locales
+
   /** @type {string} **/
   #localesFolder
 
@@ -86,6 +89,22 @@ class ComponentBuild {
    */
   set javascriptFile (value) {
     this.#javascriptFile = value
+  }
+
+  /**
+   * Get Locales
+   * @return {Object}
+   */
+  get locales () {
+    return this.#locales
+  }
+
+  /**
+   * Set Locales
+   * @param {Object} value
+   */
+  set locales (value) {
+    this.#locales = value
   }
 
   /**